pyproject.toml
[build-system]
requires = ["setuptools>=76", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "authed"
version = "0.2.3"
description = "Agent authentication SDK for secure service-to-service communication"
authors = [{name = "Antoni Gmitruk", email = "antoni@getauthed.dev"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration :: Authentication/Directory",
]

dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "pydantic>=2.10.6",
    "cryptography>=44.0.0",
    "pyjwt>=2.10.1",
    "websockets>=15.0.1",
    "uvicorn[standard]>=0.34.0",
    "authed-mcp>=0.1.2"
]

[project.urls]
Documentation = "https://docs.getauthed.dev"
"Source code" = "https://github.com/authed-dev/authed"

[project.scripts]
authed = "client.cli.main:cli"

[tool.setuptools.packages.find]
include = ["client*"]  # This will include client and all its subpackages
exclude = ["registry*", "tests*", "integrations*"]  # Exclude registry and tests directories
